requirements.txt
sentence-transformers
uvicorn
langchain_community
ctransformers
langchain
python-box
streamlit
python-dotenv